requirements.txt
#transformers==4.32.0
transformers
torch
nltk==3.8.1
requests
langdetect==1.0.9
pandas
matplotlib
seaborn
jupyter
ipykernel
numpy<2
openai
python-dotenv
accelerate
sentencepiece
protobuf